+++ CHANGELOG.md
@@ -4,11 +4,12 @@ CHANGELOG
 v2.0.0-RC.4 (20.02.2021)
 -------------------
 
-- PHP tests use latest signatures (https://github.com/spiral/roadrunner/pull/550)
-- Endure container update to v1.0.0-RC.2 version
+- PHP tests use latest signatures (https://github.com/spiral/roadrunner/pull/550).
+- Endure container update to v1.0.0-RC.2 version.
 - Remove unneeded mutex from the `http.Workers` method.
-- Rename `checker` plugin package to `status`, remove `/v1` endpoint prefix (#557) 
+- Rename `checker` plugin package to `status`, remove `/v1` endpoint prefix (#557).
 - Add static, headers, status, gzip plugins to the `main.go`.
+- Fix workers pool behavior -> idle_ttl, ttl, max_memory are soft errors and exec_ttl is hard error.
 
 v2.0.0-RC.3 (17.02.2021)
 -------------------